Improve webpack config
Add sass, jade, angular template loader, html-dist plugin, and ng annotate with angular and react flags
