+++ index.js
@@ -1,5 +1,5 @@
 module.exports = {
   html_dist_defaults: require('./html-dist.conf'),
   karma_defaults: require('./karma.conf'),
-  webpack_defaults: require('./webpack.conf'),
+  webpack_get_defaults: require('./webpack_get_defaults'),
 }
