+++ package.json
@@ -14,14 +14,22 @@
   "author": "Polyconseil",
   "license": "MIT",
   "dependencies": {
+  },
+  "files": [
+    ".eslintrc",
+    "html-dist.conf.js",
+    "js.mk",
+    "karma.conf.js",
+    "package.json",
+    "webpack.conf.js",
+    "README.md"
+  ],
+  "devDependencies": {
     "babel-loader": "6.2.2",
     "babel-polyfill": "6.5.0",
     "babel-preset-es2015": "6.5.0",
     "concurrently": "1.0.0",
-    "css-loader": "0.23.1",
     "eslint": "2.0.0",
-    "file-loader": "0.8.5",
-    "html-dist": "0.2.2",
     "jasmine-core": "2.4.1",
     "karma-coverage": "0.5.3",
     "karma-jasmine": "0.3.7",
@@ -36,17 +44,20 @@
     "live-server": "0.9.1",
     "phantomjs-prebuilt": "2.1.4",
     "standard": "6.0.4",
+    "file-loader": "0.8.5",
+    "css-loader": "0.23.1",
+    "sass-loader": "3.1.2",
     "style-loader": "0.13.0",
     "url-loader": "0.5.7",
-    "webpack": "1.12.13"
-  },
-  "files": [
-    ".eslintrc",
-    "html-dist.conf.js",
-    "js.mk",
-    "karma.conf.js",
-    "package.json",
-    "webpack.conf.js",
-    "README.md"
-  ]
+    "webpack": "1.12.13",
+    "html-loader": "0.4.3",
+    "html-webpack-plugin": "2.9.0",
+    "jade": "1.11.0",
+    "jade-html-loader": "0.0.3",
+    "jade-loader": "0.8.0",
+    "ng-annotate-webpack-plugin": "0.1.2",
+    "ngtemplate-loader": "1.3.1",
+    "node-sass": "3.4.2",
+    "postcss-loader": "0.8.1"
+  }
 }
